phoebe_committee_file_data_and_implementation
