+++ src/app/board/page.tsx
@@ -1,9 +1,11 @@
 import MemberCards from "@/components/board/MemberCards";
+import CommitteeCards from "@/components/board/CommitteeCards";
 
 const Board = () => {
   return (
     <div>
       <MemberCards />
+      <CommitteeCards />
     </div>
   );
 };
