+++ src/components/board/CommitteeCard.tsx
@@ -1,0 +1,54 @@
+export const resComInfo = [
+  {
+    name: "Sapesh Ravichandren",
+    email: "example@email.com",
+  },
+  {
+    name: "Angela Victoria Martinez",
+    email: "example@email.com",
+  },
+  {
+    name: "Gauransh Gulati",
+    email: "example@email.com",
+  },
+  {
+    name: "Niki Azizi",
+    email: "example@email.com",
+  },
+  {
+    name: "Kenda Sinjab",
+    email: "example@email.com",
+  },
+  {
+    name: "Diana Xue",
+    email: "example@email.com",
+  },
+  {
+    name: "Jiani Song",
+    email: "example@email.com",
+  },
+  {
+    name: "Priya Jacob",
+    email: "example@email.com",
+  },
+  {
+    name: "Sadia Uddin",
+    email: "example@email.com",
+  },
+  {
+    name: "Aryan Iyyappan",
+    email: "example@email.com",
+  },
+  {
+    name: "Amber Spatare",
+    email: "example@email.com",
+  },
+  {
+    name: "Amyra El Khatib",
+    email: "example@email.com",
+  },
+  {
+    name: "Bella Chana",
+    email: "example@email.com",
+  },
+];
